Asio+Beast-based endpoint server (#274)
* Update dependency versions

* boost::beast and asio-based asynchronous endpoint implementation

* Address review comments

* Remove Pistache as it is no longer used

* Fix TSan-detected data races in distributed tests

* Turn off MPI all-to-all disttests, see <https://github.com/faasm/faabric/issues/275>

* Run clang-format-13

+++ src/transport/MessageEndpoint.cpp
@@ -122,8 +122,9 @@ zmq::socket_t socketFactory(zmq::socket_type socketType,
                     break;
                 }
                 default: {
-                    SPDLOG_ERROR(
-                      "Invalid bind socket type {} ({})", socketType, address);
+                    SPDLOG_ERROR("Invalid bind socket type {} ({})",
+                                 static_cast<int>(socketType),
+                                 address);
                     throw std::runtime_error(
                       "Binding with invalid socket type");
                 }
@@ -176,7 +177,7 @@ zmq::socket_t socketFactory(zmq::socket_type socketType,
                 }
                 default: {
                     SPDLOG_ERROR("Invalid connect socket type {} ({})",
-                                 socketType,
+                                 static_cast<int>(socketType),
                                  address);
                     throw std::runtime_error(
                       "Connecting with unrecognized socket type");
